Introduce a custom struct to store the difference and save it in ctx
Added a custom type that holds the difference data for every stage. This
type will be shared through contexts and will be used in recording the
logs as well.
